+++ src/state.c
@@ -55,7 +55,7 @@ void free_aerc_message(struct aerc_message *msg) {
 	if (!msg) return;
 	free_flat_list(msg->flags);
 	if (msg->headers) {
-		for (int i = 0; i < msg->headers->length; ++i) {
+		for (size_t i = 0; i < msg->headers->length; ++i) {
 			struct email_header *header = msg->headers->items[i];
 			free(header->key);
 			free(header->value);
@@ -66,7 +66,7 @@ void free_aerc_message(struct aerc_message *msg) {
 }
 
 const char *get_message_header(struct aerc_message *msg, char *key) {
-	for (int i = 0; i < msg->headers->length; ++i) {
+	for (size_t i = 0; i < msg->headers->length; ++i) {
 		struct email_header *header = msg->headers->items[i];
 		if (strcmp(header->key, key) == 0) {
 			return header->value;
@@ -77,7 +77,7 @@ const char *get_message_header(struct aerc_message *msg, char *key) {
 
 bool get_mailbox_flag(struct aerc_mailbox *mbox, char *flag) {
 	if (!mbox->flags) return false;
-	for (int i = 0; i < mbox->flags->length; ++i) {
+	for (size_t i = 0; i < mbox->flags->length; ++i) {
 		const char *_flag = mbox->flags->items[i];
 		if (strcmp(flag, _flag) == 0) {
 			return true;
@@ -88,7 +88,7 @@ bool get_mailbox_flag(struct aerc_mailbox *mbox, char *flag) {
 
 bool get_message_flag(struct aerc_message *msg, char *flag) {
 	if (!msg->flags) return false;
-	for (int i = 0; i < msg->flags->length; ++i) {
+	for (size_t i = 0; i < msg->flags->length; ++i) {
 		const char *_flag = msg->flags->items[i];
 		if (strcmp(flag, _flag) == 0) {
 			return true;
@@ -98,7 +98,7 @@ bool get_message_flag(struct aerc_message *msg, char *flag) {
 }
 
 struct account_config *config_for_account(const char *name) {
-	for (int i = 0; i < config->accounts->length; ++i) {
+	for (size_t i = 0; i < config->accounts->length; ++i) {
 		struct account_config *c = config->accounts->items[i];
 		if (strcmp(c->name, name) == 0) {
 			return c;
